requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.3
distlib==0.3.6
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.87.0
fastapi-localization==0.0a2.dev3
fastapi-sessions==0.3.2
filelock==3.8.0
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.31
platformdirs==2.5.4
protobuf==3.20.1
pycparser==2.21
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
typing_extensions==4.4.0
uvicorn==0.19.0
virtualenv==20.16.7
watchfiles==0.18.1
websockets==10.4
